scene-parser/requirements.txt
flask
flask-cors
torch
transformers
llama-cpp-python